Adds configurable aggregationType support
Much like commit 0cc664 in the python whisper project,
this commit introduces support for specifying the aggregation type
(one of 'sum', 'max', 'min', 'last' or 'average') when propagating
values into the lower precision archives.

This commit also bumps the package number as it introduces a breaking
change to the create interface by making the xFilesFactor and new
aggregationMethod arguments optional (defaulting to 0.5 and average
respectively). Following semver.org recommendations that means a bump
to the middle value :)

Signed-off-by: ciaranj <ciaranj@gmail.com>