unixTime() will now round up, rather than down.
This change may be un-acceptable :(.  If Hoard.js is used within the same
process as statsd (as for example a back-end) then it is possible that hoard
will refuse updates from statsD's flush event as the timestamp that
statsd generates will appear to be from the future.

This is because Hoard.js was using:

  parseInt( new Date().getTime() / 1000 )

But statsd is using :

  Math.round( new Date().getTime() / 1000 )

To get their idea of 'now'.  Most of the time this will be ok, but if
a the result of the division contains a fractional part greater than or
equal to .5 then statsd will round-up to the next nearest integer value
but Hoard was truncating (discarding?) the fractional part, meaning
it would be (as far as it was concerned) timestamps from the future
when update was called sufficiently quickly after the flush.

Signed-off-by: ciaranj <ciaranj@gmail.com>